13.Script변경점 감지 TestCode / Map 변경점 / CRKEngine 변경점/CScriptReLoad Class 생성/FileManager 변경점 https://rokaisback.tistory.com/57

+++ Project/RokaSTL_Lib/MemoryPool.cpp
@@ -82,7 +82,7 @@ void MemoryPool::Release()
 
 void* MemoryPool::operator new(std::size_t _size)
 {
-	int size = AssignSize(_size);
+	size_t size = AssignSize(_size);
 	if (size == (int)EMemErrType::POOLSIZEOVER)
 	{
 		return malloc(_size); // 단일 메모리 할당.
